Drop `assert-image-size` threshold
Builds [are failing](https://github.com/hazelcast/hazelcast-docker/actions/runs/20253701874/job/58151102188) as the `slim` EE builds have an efficiency of 87%.

This is because the `slim` builds are much smaller than the "full" ones tested against, and the "wasted" space of the updated DNF metadata files is more significant.

This is a _characteristic_ of the newer base image, rather than a _defect_ in our Docker image - although a workaround _could_ be to hack away and it's internal metadata store - the primary usecase for this action is to find mistakes _we_ have made.